src/app_customermgmt/requirements.txt
Django==2.1.2
djangorestframework==3.9.0
psycopg2-binary==2.7.5
confluent-kafka==0.11.6
